Investigate a time zone discrepancy bug
When the server and client are in different time zones, there occurs a
bug where LogWeightWidget is matching its date argument to more than one
record. The bug is replicable but with slightly different procedures,
which I suspect is because of the dependence on how node is getting the
local timezone, which is likely different depending on operating system.
Further investigation is needed, but for now I will set the server
timezone to the same as where I am.