Fix time zone discrepancy bug
The user's time zone information should be passed to the server so it
knows how to format it back to the user. The bug occured when the client
and server had differing time zones, which wasn't being communicated
between client and server before. Internally, log-weights uses
.startOf('day') and endOf('day') extensively, which is a problem if the time zones are
different since that could mean different absolute times for the start
and end so when the client gets back results, the start/end would be in
terms of the server's time zone - leading to discrepancy. This was fixed
by having the client send its UTC offset (time zone info) via
moment.format() and having the server parse that info to keep everything
consistent.